pyproject.toml
[project]
name = "create-carrier-card"
version = "0.1.0"
description = ""
authors = [
    {name = "Denis Klyachin",email = "yithling@proton.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "aiogram (>=3.18.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "2captcha-python (>=1.5.1,<2.0.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "create_carrier_card", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
